Conflicts:
	include/onepoll.php

+++ include/onepoll.php
@@ -15,11 +15,11 @@ function RemoveReply($subject) {
 function onepoll_run(&$argv, &$argc){
 	global $a, $db;
 
-	if(is_null($a)) {
+	if (is_null($a)) {
 		$a = new App;
 	}
 
-	if(is_null($db)) {
+	if (is_null($db)) {
 		@include(".htconfig.php");
 		require_once("include/dba.php");
 		$db = new dba($db_host, $db_user, $db_pass, $db_data);
@@ -48,21 +48,25 @@ function onepoll_run(&$argv, &$argc){
 	$force      = false;
 	$restart    = false;
 
-	if(($argc > 1) && (intval($argv[1])))
+	if (($argc > 1) && (intval($argv[1]))) {
 		$contact_id = intval($argv[1]);
+	}
 
-	if(($argc > 2) && ($argv[2] == "force"))
+	if (($argc > 2) && ($argv[2] == "force")) {
 		$force = true;
+	}
 
-	if(! $contact_id) {
+	if (! $contact_id) {
 		logger('onepoll: no contact');
 		return;
 	}
 
 	// Don't check this stuff if the function is called by the poller
-	if (App::callstack() != "poller_run")
-		if (App::is_already_running('onepoll'.$contact_id, '', 540))
+	if (App::callstack() != "poller_run") {
+		if (App::is_already_running('onepoll'.$contact_id, '', 540)) {
 			return;
+		}
+	}
 
 	$d = datetime_convert();
 
@@ -83,8 +87,9 @@ function onepoll_run(&$argv, &$argc){
 		intval($contact_id)
 	);
 
-	if(! count($contacts))
+	if (! count($contacts)) {
 		return;
+	}
 
 	$contact = $contacts[0];
 
@@ -94,9 +99,11 @@ function onepoll_run(&$argv, &$argc){
 			where `cid` = %d and updated > UTC_TIMESTAMP() - INTERVAL 1 DAY",
 			intval($contact['id'])
 		);
-		if (dbm::is_result($r))
-			if (!$r[0]['total'])
+		if (dbm::is_result($r)) {
+			if (!$r[0]['total']) {
 				poco_load($contact['id'],$importer_uid,0,$contact['poco']);
+			}
+		}
 	}
 
 	/// @TODO Check why we don't poll the Diaspora feed at the moment (some guid problem in the items?)
@@ -127,24 +134,24 @@ function onepoll_run(&$argv, &$argc){
 
 	$t = $contact['last-update'];
 
-	if($contact['subhub']) {
+	if ($contact['subhub']) {
 		$poll_interval = get_config('system','pushpoll_frequency');
 		$contact['priority'] = (($poll_interval !== false) ? intval($poll_interval) : 3);
 		$hub_update = false;
 
-		if(datetime_convert('UTC','UTC', 'now') > datetime_convert('UTC','UTC', $t . " + 1 day"))
-				$hub_update = true;
-	}
-	else
+		if (datetime_convert('UTC','UTC', 'now') > datetime_convert('UTC','UTC', $t . " + 1 day")) {
+			$hub_update = true;
+		}
+	} else {
 		$hub_update = false;
-
+	}
 
 	$importer_uid = $contact['uid'];
 
 	$r = q("SELECT `contact`.*, `user`.`page-flags` FROM `contact` INNER JOIN `user` on `contact`.`uid` = `user`.`uid` WHERE `user`.`uid` = %d AND `contact`.`self` = 1 LIMIT 1",
 		intval($importer_uid)
 	);
-	if (! dbm::is_result($r)) {
+	if (!dbm::is_result($r)) {
 		return;
 	}
 
@@ -158,7 +165,7 @@ function onepoll_run(&$argv, &$argc){
 	);
 
 	// Update the contact entry
-	if(($contact['network'] === NETWORK_OSTATUS) || ($contact['network'] === NETWORK_DIASPORA) || ($contact['network'] === NETWORK_DFRN)) {
+	if (($contact['network'] === NETWORK_OSTATUS) || ($contact['network'] === NETWORK_DIASPORA) || ($contact['network'] === NETWORK_DFRN)) {
 		if (!poco_reachable($contact['url'])) {
 			logger("Skipping probably dead contact ".$contact['url']);
 			return;
@@ -167,40 +174,50 @@ function onepoll_run(&$argv, &$argc){
 		if (!update_contact($contact["id"])) {
 			mark_for_death($contact);
 			return;
-		} else
+		} else {
 			unmark_for_death($contact);
+		}
 	}
 
-	if($contact['network'] === NETWORK_DFRN) {
+	if ($contact['network'] === NETWORK_DFRN) {
 
 
 		$idtosend = $orig_id = (($contact['dfrn-id']) ? $contact['dfrn-id'] : $contact['issued-id']);
-		if(intval($contact['duplex']) && $contact['dfrn-id'])
+		if (intval($contact['duplex']) && $contact['dfrn-id']) {
 			$idtosend = '0:' . $orig_id;
-		if(intval($contact['duplex']) && $contact['issued-id'])
+		}
+		if (intval($contact['duplex']) && $contact['issued-id']) {
 			$idtosend = '1:' . $orig_id;
+		}
 
 		// they have permission to write to us. We already filtered this in the contact query.
 		$perm = 'rw';
 
 		// But this may be our first communication, so set the writable flag if it isn't set already.
 
-		if(! intval($contact['writable']))
+		if (! intval($contact['writable'])) {
 			q("update contact set writable = 1 where id = %d", intval($contact['id']));
-
+		}
 
 		$url = $contact['poll'] . '?dfrn_id=' . $idtosend
 			. '&dfrn_version=' . DFRN_PROTOCOL_VERSION
 			. '&type=data&last_update=' . $last_update
 			. '&perm=' . $perm ;
 
-		$handshake_xml = fetch_url($url);
+		$ret = z_fetch_url($url);
+
+		if ($ret['errno'] == CURLE_OPERATION_TIMEDOUT) {
+			return;
+		}
+
+		$handshake_xml = $ret['body'];
+
 		$html_code = $a->get_curl_code();
 
 		logger('onepoll: handshake with url ' . $url . ' returns xml: ' . $handshake_xml, LOGGER_DATA);
 
 
-		if((! strlen($handshake_xml)) || ($html_code >= 400) || (! $html_code)) {
+		if ((! strlen($handshake_xml)) || ($html_code >= 400) || (! $html_code)) {
 			logger("poller: $url appears to be dead - marking for death ");
 
 			// dead connection - might be a transient event, or this might
@@ -219,7 +236,7 @@ function onepoll_run(&$argv, &$argc){
 			return;
 		}
 
-		if(! strstr($handshake_xml,'<')) {
+		if (! strstr($handshake_xml,'<')) {
 			logger('poller: response from ' . $url . ' did not contain XML.');
 
 			mark_for_death($contact);
@@ -235,7 +252,7 @@ function onepoll_run(&$argv, &$argc){
 
 		$res = parse_xml_string($handshake_xml);
 
-		if(intval($res->status) == 1) {
+		if (intval($res->status) == 1) {
 			logger("poller: $url replied status 1 - marking for death ");
 
 			// we may not be friends anymore. Will keep trying for one month.
@@ -248,18 +265,16 @@ function onepoll_run(&$argv, &$argc){
 				intval($contact['id'])
 			);
 			mark_for_death($contact);
-		}
-		else {
-			if ($contact['term-date'] > NULL_DATE) {
-				logger("poller: $url back from the dead - removing mark for death");
-				unmark_for_death($contact);
-			}
+		} elseif ($contact['term-date'] > NULL_DATE) {
+			logger("poller: $url back from the dead - removing mark for death");
+			unmark_for_death($contact);
 		}
 
-		if((intval($res->status) != 0) || (! strlen($res->challenge)) || (! strlen($res->dfrn_id)))
+		if ((intval($res->status) != 0) || (! strlen($res->challenge)) || (! strlen($res->dfrn_id))) {
 			return;
+		}
 
-		if(((float) $res->dfrn_version > 2.21) && ($contact['poco'] == '')) {
+		if (((float) $res->dfrn_version > 2.21) && ($contact['poco'] == '')) {
 			q("update contact set poco = '%s' where id = %d",
 				dbesc(str_replace('/profile/','/poco/', $contact['url'])),
 				intval($contact['id'])
@@ -273,21 +288,21 @@ function onepoll_run(&$argv, &$argc){
 
 		$final_dfrn_id = '';
 
-		if(($contact['duplex']) && strlen($contact['prvkey'])) {
+		if (($contact['duplex']) && strlen($contact['prvkey'])) {
 			openssl_private_decrypt($sent_dfrn_id,$final_dfrn_id,$contact['prvkey']);
 			openssl_private_decrypt($challenge,$postvars['challenge'],$contact['prvkey']);
-		}
-		else {
+		} else {
 			openssl_public_decrypt($sent_dfrn_id,$final_dfrn_id,$contact['pubkey']);
 			openssl_public_decrypt($challenge,$postvars['challenge'],$contact['pubkey']);
 		}
 
 		$final_dfrn_id = substr($final_dfrn_id, 0, strpos($final_dfrn_id, '.'));
 
-		if(strpos($final_dfrn_id,':') == 1)
+		if (strpos($final_dfrn_id,':') == 1) {
 			$final_dfrn_id = substr($final_dfrn_id,2);
+		}
 
-		if($final_dfrn_id != $orig_id) {
+		if ($final_dfrn_id != $orig_id) {
 			logger('poller: ID did not decode: ' . $contact['id'] . ' orig: ' . $orig_id . ' final: ' . $final_dfrn_id);
 			// did not decode properly - cannot trust this site
 			return;
@@ -299,10 +314,9 @@ function onepoll_run(&$argv, &$argc){
 
 		$xml = post_url($contact['poll'],$postvars);
 
-	}
-	elseif(($contact['network'] === NETWORK_OSTATUS)
+	} elseif (($contact['network'] === NETWORK_OSTATUS)
 		|| ($contact['network'] === NETWORK_DIASPORA)
-		|| ($contact['network'] === NETWORK_FEED) ) {
+		|| ($contact['network'] === NETWORK_FEED)) {
 
 		// Upgrading DB fields from an older Friendica version
 		// Will only do this once per notify-enabled OStatus contact
@@ -311,10 +325,11 @@ function onepoll_run(&$argv, &$argc){
 		$stat_writeable = ((($contact['notify']) && ($contact['rel'] == CONTACT_IS_FOLLOWER || $contact['rel'] == CONTACT_IS_FRIEND)) ? 1 : 0);
 
 		// Contacts from OStatus are always writable
-		if($contact['network'] === NETWORK_OSTATUS)
+		if ($contact['network'] === NETWORK_OSTATUS) {
 			$stat_writeable = 1;
+		}
 
-		if($stat_writeable != $contact['writable']) {
+		if ($stat_writeable != $contact['writable']) {
 			q("UPDATE `contact` SET `writable` = %d WHERE `id` = %d",
 				intval($stat_writeable),
 				intval($contact['id'])
@@ -323,19 +338,26 @@ function onepoll_run(&$argv, &$argc){
 
 		// Are we allowed to import from this person?
 
-		if($contact['rel'] == CONTACT_IS_FOLLOWER || $contact['blocked'] || $contact['readonly'])
+		if ($contact['rel'] == CONTACT_IS_FOLLOWER || $contact['blocked'] || $contact['readonly']) {
 			return;
+		}
 
 		$cookiejar = tempnam(get_temppath(), 'cookiejar-onepoll-');
-		$xml = fetch_url($contact['poll'], false, $redirects, 0, Null, $cookiejar);
+		$ret = z_fetch_url($contact['poll'], false, $redirects, array('cookiejar' => $cookiejar));
+
+		if ($ret['errno'] == CURLE_OPERATION_TIMEDOUT) {
+			return;
+		}
+
+		$xml = $ret['body'];
+
 		unlink($cookiejar);
-	}
-	elseif($contact['network'] === NETWORK_MAIL || $contact['network'] === NETWORK_MAIL2) {
+	} elseif ($contact['network'] === NETWORK_MAIL || $contact['network'] === NETWORK_MAIL2) {
 
 		logger("Mail: Fetching", LOGGER_DEBUG);
 
 		$mail_disabled = ((function_exists('imap_open') && (! get_config('system','imap_disabled'))) ? 0 : 1);
-		if($mail_disabled)
+		if ($mail_disabled)
 			return;
 
 		logger("Mail: Enabled", LOGGER_DEBUG);
@@ -347,38 +369,38 @@ function onepoll_run(&$argv, &$argc){
 		$mailconf = q("SELECT * FROM `mailacct` WHERE `server` != '' AND `uid` = %d LIMIT 1",
 			intval($importer_uid)
 		);
-		if(count($x) && count($mailconf)) {
+		if (count($x) && count($mailconf)) {
 		    $mailbox = construct_mailbox_name($mailconf[0]);
 			$password = '';
 			openssl_private_decrypt(hex2bin($mailconf[0]['pass']),$password,$x[0]['prvkey']);
 			$mbox = email_connect($mailbox,$mailconf[0]['user'],$password);
 			unset($password);
 			logger("Mail: Connect to " . $mailconf[0]['user']);
-			if($mbox) {
+			if ($mbox) {
 				q("UPDATE `mailacct` SET `last_check` = '%s' WHERE `id` = %d AND `uid` = %d",
 					dbesc(datetime_convert()),
 					intval($mailconf[0]['id']),
 					intval($importer_uid)
 				);
 				logger("Mail: Connected to " . $mailconf[0]['user']);
-			} else
+			} else {
 				logger("Mail: Connection error ".$mailconf[0]['user']." ".print_r(imap_errors(), true));
+			}
 		}
-		if($mbox) {
+		if ($mbox) {
 
 			$msgs = email_poll($mbox,$contact['addr']);
 
-			if(count($msgs)) {
+			if (count($msgs)) {
 				logger("Mail: Parsing ".count($msgs)." mails for ".$mailconf[0]['user'], LOGGER_DEBUG);
 
 				$metas = email_msg_meta($mbox,implode(',',$msgs));
-				if(count($metas) != count($msgs)) {
+				if (count($metas) != count($msgs)) {
 					logger("onepoll: for " . $mailconf[0]['user'] . " there are ". count($msgs) . " messages but received " . count($metas) . " metas", LOGGER_DEBUG);
-				}
-				else {
+				} else {
 					$msgs = array_combine($msgs, $metas);
 
-					foreach($msgs as $msg_uid => $meta) {
+					foreach ($msgs as $msg_uid => $meta) {
 						logger("Mail: Parsing mail ".$msg_uid, LOGGER_DATA);
 
 						$datarray = array();
@@ -400,7 +422,7 @@ function onepoll_run(&$argv, &$argc){
 
 							// Only delete when mails aren't automatically moved or deleted
 							if (($mailconf[0]['action'] != 1) AND ($mailconf[0]['action'] != 3))
-								if($meta->deleted && ! $r[0]['deleted']) {
+								if ($meta->deleted && ! $r[0]['deleted']) {
 									q("UPDATE `item` SET `deleted` = 1, `changed` = '%s' WHERE `id` = %d",
 										dbesc(datetime_convert()),
 										intval($r[0]['id'])
@@ -434,13 +456,13 @@ function onepoll_run(&$argv, &$argc){
 
 	//					$raw_refs = ((x($headers,'references')) ? str_replace("\t",'',$headers['references']) : '');
 						$raw_refs = ((property_exists($meta,'references')) ? str_replace("\t",'',$meta->references) : '');
-						if(! trim($raw_refs))
+						if (! trim($raw_refs))
 							$raw_refs = ((property_exists($meta,'in_reply_to')) ? str_replace("\t",'',$meta->in_reply_to) : '');
 						$raw_refs = trim($raw_refs);  // Don't allow a blank reference in $refs_arr
 
-						if($raw_refs) {
+						if ($raw_refs) {
 							$refs_arr = explode(' ', $raw_refs);
-							if(count($refs_arr)) {
+							if (count($refs_arr)) {
 								for($x = 0; $x < count($refs_arr); $x ++)
 									$refs_arr[$x] = "'" . msgid2iri(str_replace(array('<','>',' '),array('','',''),dbesc($refs_arr[$x]))) . "'";
 							}
@@ -456,12 +478,13 @@ function onepoll_run(&$argv, &$argc){
 						// Decoding the header
 						$subject = imap_mime_header_decode($meta->subject);
 						$datarray['title'] = "";
-						foreach($subject as $subpart)
-							if ($subpart->charset != "default")
+						foreach ($subject as $subpart) {
+							if ($subpart->charset != "default") {
 								$datarray['title'] .= iconv($subpart->charset, 'UTF-8//IGNORE', $subpart->text);
-							else
+							} else {
 								$datarray['title'] .= $subpart->text;
-
+							}
+						}
 						$datarray['title'] = notags(trim($datarray['title']));
 
 						//$datarray['title'] = notags(trim($meta->subject));
@@ -476,7 +499,7 @@ function onepoll_run(&$argv, &$argc){
 						$datarray['title'] = RemoveReply($datarray['title']);
 
 						// If it seems to be a reply but a header couldn't be found take the last message with matching subject
-						if(!x($datarray,'parent-uri') and $reply) {
+						if (!x($datarray,'parent-uri') and $reply) {
 							$r = q("SELECT `uri` , `parent-uri` FROM `item` WHERE `title` = \"%s\" AND `uid` = %d AND `network` = '%s' ORDER BY `created` DESC LIMIT 1",
 								dbesc(protect_sprintf($datarray['title'])),
 								intval($importer_uid),
@@ -485,12 +508,12 @@ function onepoll_run(&$argv, &$argc){
 								$datarray['parent-uri'] = $r[0]['parent-uri'];
 						}
 
-						if(! x($datarray,'parent-uri'))
+						if (! x($datarray,'parent-uri'))
 							$datarray['parent-uri'] = $datarray['uri'];
 
 
 						$r = email_get_msg($mbox,$msg_uid, $reply);
-						if(! $r) {
+						if (! $r) {
 							logger("Mail: can't fetch msg ".$msg_uid." for ".$mailconf[0]['user']);
 							continue;
 						}
@@ -502,23 +525,26 @@ function onepoll_run(&$argv, &$argc){
 						// some mailing lists have the original author as 'from' - add this sender info to msg body.
 						/// @TODO Adding a gravatar for the original author would be cool
 
-						if(! stristr($meta->from,$contact['addr'])) {
+						if (! stristr($meta->from,$contact['addr'])) {
 							$from = imap_mime_header_decode($meta->from);
 							$fromdecoded = "";
-							foreach($from as $frompart)
-								if ($frompart->charset != "default")
+							foreach ($from as $frompart) {
+								if ($frompart->charset != "default") {
 									$fromdecoded .= iconv($frompart->charset, 'UTF-8//IGNORE', $frompart->text);
-								else
+								} else {
 									$fromdecoded .= $frompart->text;
+								}
+							}
 
 							$fromarr = imap_rfc822_parse_adrlist($fromdecoded, $a->get_hostname());
 
 							$frommail = $fromarr[0]->mailbox."@".$fromarr[0]->host;
 
-							if (isset($fromarr[0]->personal))
+							if (isset($fromarr[0]->personal)) {
 								$fromname = $fromarr[0]->personal;
-							else
+							} else {
 								$fromname = $frommail;
+							}
 
 							//$datarray['body'] = "[b]".t('From: ') . escape_tags($fromdecoded) . "[/b]\n\n" . $datarray['body'];
 
@@ -538,9 +564,9 @@ function onepoll_run(&$argv, &$argc){
 
 						$datarray['uid'] = $importer_uid;
 						$datarray['contact-id'] = $contact['id'];
-						if($datarray['parent-uri'] === $datarray['uri'])
+						if ($datarray['parent-uri'] === $datarray['uri'])
 							$datarray['private'] = 1;
-						if(($contact['network'] === NETWORK_MAIL) && (! get_pconfig($importer_uid,'system','allow_public_email_replies'))) {
+						if (($contact['network'] === NETWORK_MAIL) && (! get_pconfig($importer_uid,'system','allow_public_email_replies'))) {
 							$datarray['private'] = 1;
 							$datarray['allow_cid'] = '<' . $contact['id'] . '>';
 						}
@@ -574,24 +600,24 @@ function onepoll_run(&$argv, &$argc){
 						}
 					}
 				}
-			} else
+			} else {
 				logger("Mail: no mails for ".$mailconf[0]['user']);
+			}
 
 			logger("Mail: closing connection for ".$mailconf[0]['user']);
 			imap_close($mbox);
 		}
-	}
-	elseif($contact['network'] === NETWORK_FACEBOOK) {
+	} elseif ($contact['network'] === NETWORK_FACEBOOK) {
 		// This is picked up by the Facebook plugin on a cron hook.
 		// Ignored here.
-	} elseif($contact['network'] === NETWORK_PUMPIO) {
+	} elseif ($contact['network'] === NETWORK_PUMPIO) {
 		// This is picked up by the pump.io plugin on a cron hook.
 		// Ignored here.
 	}
 
-	if($xml) {
+	if ($xml) {
 		logger('poller: received xml : ' . $xml, LOGGER_DATA);
-		if(! strstr($xml,'<')) {
+		if (! strstr($xml,'<')) {
 			logger('poller: post_handshake: response from ' . $url . ' did not contain XML.');
 			$r = q("UPDATE `contact` SET `last-update` = '%s',  `failure_update` = '%s' WHERE `id` = %d",
 				dbesc(datetime_convert()),
@@ -611,10 +637,10 @@ function onepoll_run(&$argv, &$argc){
 		consume_feed($xml,$importer,$contact,$hub,1,2);
 
 		$hubmode = 'subscribe';
-		if($contact['network'] === NETWORK_DFRN || $contact['blocked'] || $contact['readonly'])
+		if ($contact['network'] === NETWORK_DFRN || $contact['blocked'] || $contact['readonly'])
 			$hubmode = 'unsubscribe';
 
-		if(($contact['network'] === NETWORK_OSTATUS ||  $contact['network'] == NETWORK_FEED) && (! $contact['hub-verify']))
+		if (($contact['network'] === NETWORK_OSTATUS ||  $contact['network'] == NETWORK_FEED) && (! $contact['hub-verify']))
 			$hub_update = true;
 
 		if ($force)
@@ -622,14 +648,15 @@ function onepoll_run(&$argv, &$argc){
 
 		logger("Contact ".$contact['id']." returned hub: ".$hub." Network: ".$contact['network']." Relation: ".$contact['rel']." Update: ".$hub_update);
 
-		if((strlen($hub)) && ($hub_update) && (($contact['rel'] != CONTACT_IS_FOLLOWER) || $contact['network'] == NETWORK_FEED) ) {
+		if ((strlen($hub)) && ($hub_update) && (($contact['rel'] != CONTACT_IS_FOLLOWER) || $contact['network'] == NETWORK_FEED) ) {
 			logger('poller: hub ' . $hubmode . ' : ' . $hub . ' contact name : ' . $contact['name'] . ' local user : ' . $importer['name']);
 			$hubs = explode(',', $hub);
-			if(count($hubs)) {
-				foreach($hubs as $h) {
+			if (count($hubs)) {
+				foreach ($hubs as $h) {
 					$h = trim($h);
-					if(! strlen($h))
+					if (! strlen($h)) {
 						continue;
+					}
 					subscribe_to_hub($h,$importer,$contact,$hubmode);
 				}
 			}
@@ -671,7 +698,7 @@ function onepoll_run(&$argv, &$argc){
 	return;
 }
 
-if (array_search(__file__,get_included_files())===0){
+if (array_search(__file__,get_included_files())===0) {
 	onepoll_run($_SERVER["argv"],$_SERVER["argc"]);
 	killme();
 }
